Tests/LibWeb/Text/input/selection-over-multiple-code-units.html

<!DOCTYPE html>
<script src="include.js"></script>
😭foobar😭
<script>
test(() => {
    internals.mouseDown(55, 20);

    println('--- hit test ---');
    let hit = internals.hitTest(55, 20);
    printElement(hit.node);
    println(`index: ${hit.indexInNode}`);
    printElement(hit.node.parentNode);

    internals.movePointerTo(110, 20);

    println('--- hit test ---');
    hit = internals.hitTest(110, 20);
    printElement(hit.node);
    println(`index: ${hit.indexInNode}`);
    printElement(hit.node.parentNode);

    const activeRange = window.getSelection().getRangeAt(0);
    println('');
    printElement(activeRange.startContainer);
    println(activeRange.startOffset);
    printElement(activeRange.endContainer);
    println(activeRange.endOffset);
});
</script>
