Tests/LibWeb/Text/expected/selection-over-multiple-code-units.txt

--- hit test ---
<#text>
index: 5
<BODY>
--- hit test ---
<BODY>
index: 0
<HTML>

<#text>
5
<#text>
10
